# docs/swagger.json
{
  "swagger": "2.0",
  "basePath": "/ispyb/api/v1",
  "paths": {
...
    "/samples": {
      "post": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Sample"
            }
          }
        },
        "summary": "Adds a new sample item",
        "operationId": "post_sample",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Sample"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Returns all sample items",
        "operationId": "get_sample",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Check which methods are allowed",
        "description": "Use this method if you need to know what operations are allowed to be\nperformed on this endpoint, e.g. to decide wether to display a button\nin your UI.\n\nThe list of allowed methods is provided in `Allow` response header.",
        "operationId": "options_sample",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      }
    },
    "/samples/crystals": {
      "post": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Crystal"
            }
          }
        },
        "summary": "Adds a new crystal item",
        "operationId": "post_crystals",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Crystal"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Returns all crystal items",
        "operationId": "get_crystals_by_query",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Check which methods are allowed",
        "description": "Use this method if you need to know what operations are allowed to be\nperformed on this endpoint, e.g. to decide wether to display a button\nin your UI.\n\nThe list of allowed methods is provided in `Allow` response header.",
        "operationId": "options_crystals",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      }
    },
    "/samples/crystals/{crystal_id}": {
      "parameters": [
        {
          "in": "path",
          "description": "Crystal id (integer)",
          "name": "crystal_id",
          "required": true,
          "type": "integer"
        }
      ],
      "put": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Crystal"
            }
          },
          "404": {
            "description": "Crystal not found."
          }
        },
        "summary": "Fully updates crystal with crystal_id",
        "operationId": "put_crystal_by_id",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Crystal"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "options": {
        "responses": {
          "404": {
            "description": "Crystal not found."
          }
        },
        "summary": "Check which methods are allowed",
        "description": "Use this method if you need to know what operations are allowed to be\nperformed on this endpoint, e.g. to decide wether to display a button\nin your UI.\n\nThe list of allowed methods is provided in `Allow` response header.",
        "operationId": "options_crystal_by_id",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Crystal"
            }
          },
          "404": {
            "description": "Crystal not found."
          }
        },
        "summary": "Returns a crystal by crystalId",
        "description": "crystal_id should be an integer",
        "operationId": "get_crystal_by_id",
        "parameters": [
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "delete": {
        "responses": {
          "404": {
            "description": "Crystal not found."
          }
        },
        "summary": "Deletes a crystal by crystalId",
        "operationId": "delete_crystal_by_id",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "patch": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Crystal"
            }
          },
          "404": {
            "description": "Crystal not found."
          }
        },
        "summary": "Partially updates crystal with id crystalId",
        "operationId": "patch_crystal_by_id",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Crystal"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      }
    },
    "/samples/diffraction_plans": {
      "post": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DiffractionPlan"
            }
          }
        },
        "summary": "Adds a new diffraction_plan",
        "operationId": "post_diffraction_plans",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/DiffractionPlan"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Returns list of diffraction_plans",
        "operationId": "get_diffraction_plans",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Check which methods are allowed",
        "description": "Use this method if you need to know what operations are allowed to be\nperformed on this endpoint, e.g. to decide wether to display a button\nin your UI.\n\nThe list of allowed methods is provided in `Allow` response header.",
        "operationId": "options_diffraction_plans",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      }
    },
    "/samples/diffraction_plans/{diffraction_plan_id}": {
      "parameters": [
        {
          "in": "path",
          "description": "diffraction_plan id (integer)",
          "name": "diffraction_plan_id",
          "required": true,
          "type": "integer"
        }
      ],
      "put": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DiffractionPlan"
            }
          },
          "404": {
            "description": "diffraction_plan not found."
          }
        },
        "summary": "Fully updates diffraction_plan with diffraction_plan_id",
        "operationId": "put_diffraction_plan_by_id",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/DiffractionPlan"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "options": {
        "responses": {
          "404": {
            "description": "diffraction_plan not found."
          }
        },
        "summary": "Check which methods are allowed",
        "description": "Use this method if you need to know what operations are allowed to be\nperformed on this endpoint, e.g. to decide wether to display a button\nin your UI.\n\nThe list of allowed methods is provided in `Allow` response header.",
        "operationId": "options_diffraction_plan_by_id",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DiffractionPlan"
            }
          },
          "404": {
            "description": "diffraction_plan not found."
          }
        },
        "summary": "Returns a diffraction_plan by diffraction_planId",
        "description": "diffraction_plan_id should be an integer",
        "operationId": "get_diffraction_plan_by_id",
        "parameters": [
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "delete": {
        "responses": {
          "404": {
            "description": "diffraction_plan not found."
          }
        },
        "summary": "Deletes a diffraction_plan by diffraction_planId",
        "operationId": "delete_diffraction_plan_by_id",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "patch": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DiffractionPlan"
            }
          },
          "404": {
            "description": "diffraction_plan not found."
          }
        },
        "summary": "Partially updates diffraction_plan with id diffraction_planId",
        "operationId": "patch_diffraction_plan_by_id",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/DiffractionPlan"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      }
    },
    "/samples/proteins": {
      "post": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          }
        },
        "summary": "Adds a new protein item",
        "operationId": "post_proteins",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Returns all protein items",
        "operationId": "get_proteins_by_request",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Check which methods are allowed",
        "description": "Use this method if you need to know what operations are allowed to be\nperformed on this endpoint, e.g. to decide wether to display a button\nin your UI.\n\nThe list of allowed methods is provided in `Allow` response header.",
        "operationId": "options_proteins",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      }
    },
    "/samples/proteins/{protein_id}": {
      "parameters": [
        {
          "in": "path",
          "description": "protein id (integer)",
          "name": "protein_id",
          "required": true,
          "type": "integer"
        }
      ],
      "put": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          },
          "404": {
            "description": "protein not found."
          }
        },
        "summary": "Fully updates protein with proteinId",
        "operationId": "put_protein_by_id",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "options": {
        "responses": {
          "404": {
            "description": "protein not found."
          }
        },
        "summary": "Check which methods are allowed",
        "description": "Use this method if you need to know what operations are allowed to be\nperformed on this endpoint, e.g. to decide wether to display a button\nin your UI.\n\nThe list of allowed methods is provided in `Allow` response header.",
        "operationId": "options_protein_by_id",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          },
          "404": {
            "description": "protein not found."
          }
        },
        "summary": "Returns a protein by proteinId",
        "description": "protein_id should be an integer",
        "operationId": "get_protein_by_id",
        "parameters": [
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "delete": {
        "responses": {
          "404": {
            "description": "protein not found."
          }
        },
        "summary": "Deletes a protein by proteinId",
        "operationId": "delete_protein_by_id",
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      },
      "patch": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          },
          "404": {
            "description": "protein not found."
          }
        },
        "summary": "Partially updates protein with proteinId",
        "operationId": "patch_protein_by_id",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "tags": [
          "Samples"
        ]
      }
    },
...
  },
  "info": {
    "title": "ISPyB",
    "version": "1.0",
    "description": "ISPyB Flask rest server"
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "apikey": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
...
  "definitions": {
...
    "Crystal": {
      "required": [
        "crystalId",
        "proteinId",
        "recordTimeStamp"
      ],
      "properties": {
        "crystalId": {
          "type": "integer",
          "description": ""
        },
        "diffractionPlanId": {
          "type": "integer",
          "description": ""
        },
        "proteinId": {
          "type": "integer",
          "description": ""
        },
        "crystalUUID": {
          "type": "string",
          "description": ""
        },
        "name": {
          "type": "string",
          "description": ""
        },
        "spaceGroup": {
          "type": "string",
          "description": ""
        },
        "morphology": {
          "type": "string",
          "description": ""
        },
        "color": {
          "type": "string",
          "description": ""
        },
        "size_X": {
          "type": "string",
          "description": ""
        },
        "size_Y": {
          "type": "string",
          "description": ""
        },
        "size_Z": {
          "type": "string",
          "description": ""
        },
        "cell_a": {
          "type": "string",
          "description": ""
        },
        "cell_b": {
          "type": "string",
          "description": ""
        },
        "cell_c": {
          "type": "string",
          "description": ""
        },
        "cell_alpha": {
          "type": "string",
          "description": ""
        },
        "cell_beta": {
          "type": "string",
          "description": ""
        },
        "cell_gamma": {
          "type": "string",
          "description": ""
        },
        "comments": {
          "type": "string",
          "description": ""
        },
        "pdbFileName": {
          "type": "string",
          "description": "pdb file name"
        },
        "pdbFilePath": {
          "type": "string",
          "description": "pdb file path"
        },
        "recordTimeStamp": {
          "type": "string",
          "format": "date-time",
          "description": "Creation or last update date/time"
        },
        "abundance": {
          "type": "number",
          "description": ""
        },
        "theoreticalDensity": {
          "type": "number",
          "description": ""
        }
      },
      "type": "object"
    },
...
  },
...
}